Create stable build for running simulations on PX2 and HK Computer
Changes:

Changed delays in the manager.py files in order for the launch scripts to be running in sequential order. This needs to be trimmed in the future to get a faster startup.

Broke out trajectory generator from motion planning launch file to remove a potential bug that might have existed during simulations. This should be reverted in a future patch.

Added a script for checking GPU loads. When running simulations this was at aroudn 40-45%.

Also noted that quite a high speed network connection needs to exist between Simulation environment computer and AD-EYE computer when transmitting camera feeds. With all the cameras enabled (3 of them) a bandwidth of at least 250-300Mbps is needed.

Collaborators on this patch: Maxime and Jonas Jungåker