add files for prescan experiment with real world pcd and generated mesh
Added the w10 pex experiment with new goal that is interpretable for the ADEYE platform.

The localization diverges when running on the Nvidia PX2.

added the mesh to the prescan experiment, the dae files in the actors folder as well as the maskin_mesh_cut folder in the experiments folder.

The experiment in itself is copied from the W05_KTH experiment so the functionality should be very similar but files are not renamed to match the current experiment folder name. should be fixed, sooner rather than later.

The pcd file that is running on the PX2 (linux side running adeye) is too big to be uploaded to git, therefore it is shared via a private drive link for now. A possible fix might be to divide the pcd file into several smaller. The code for meshing might be able to fix this, tests should be done to validate.

The meshing code is on other branch (hk_adeye), should be merged into this one (rcv_hk).