Added priority to SSMP endposes
Removed the use three colors to have a more flexible system.
Now the left of the ego vehicle has a malus so that stopping on the right side out of the road is preferred.
The next priority is stopping on the current lane.
Last resort is crossing the road to park on the other side.